+++ sql/pgwire/v3.go
@@ -337,13 +337,13 @@ func (c *v3Conn) handleParse(ctx context.Context, buf *readBuffer) error {
 	}
 	// The client may provide type information for (some of) the
 	// placeholders. Read this first.
-	numQArgTypes, err := buf.getInt16()
+	numQArgTypes, err := buf.getUint16()
 	if err != nil {
 		return err
 	}
 	inTypeHints := make([]oid.Oid, numQArgTypes)
 	for i := range inTypeHints {
-		typ, err := buf.getInt32()
+		typ, err := buf.getUint32()
 		if err != nil {
 			return err
 		}
@@ -491,7 +491,7 @@ func (c *v3Conn) handleBind(buf *readBuffer) error {
 	}
 
 	stmtMeta := stmt.ProtocolMeta.(preparedStatementMeta)
-	numQArgs := int16(len(stmtMeta.inTypes))
+	numQArgs := uint16(len(stmtMeta.inTypes))
 	qArgFormatCodes := make([]formatCode, numQArgs)
 	// From the docs on number of argument format codes to bind:
 	// This can be zero to indicate that there are no arguments or that the
@@ -499,7 +499,7 @@ func (c *v3Conn) handleBind(buf *readBuffer) error {
 	// specified format code is applied to all arguments; or it can equal the
 	// actual number of arguments.
 	// http://www.postgresql.org/docs/current/static/protocol-message-formats.html
-	numQArgFormatCodes, err := buf.getInt16()
+	numQArgFormatCodes, err := buf.getUint16()
 	if err != nil {
 		return err
 	}
@@ -507,7 +507,7 @@ func (c *v3Conn) handleBind(buf *readBuffer) error {
 	case 0:
 	case 1:
 		// `1` means read one code and apply it to every argument.
-		c, err := buf.getInt16()
+		c, err := buf.getUint16()
 		if err != nil {
 			return err
 		}
@@ -518,7 +518,7 @@ func (c *v3Conn) handleBind(buf *readBuffer) error {
 	case numQArgs:
 		// Read one format code for each argument and apply it to that argument.
 		for i := range qArgFormatCodes {
-			c, err := buf.getInt16()
+			c, err := buf.getUint16()
 			if err != nil {
 				return err
 			}
@@ -528,7 +528,7 @@ func (c *v3Conn) handleBind(buf *readBuffer) error {
 		return c.sendInternalError(fmt.Sprintf("wrong number of format codes specified: %d for %d arguments", numQArgFormatCodes, numQArgs))
 	}
 
-	numValues, err := buf.getInt16()
+	numValues, err := buf.getUint16()
 	if err != nil {
 		return err
 	}
@@ -537,12 +537,12 @@ func (c *v3Conn) handleBind(buf *readBuffer) error {
 	}
 	qargs := parser.QueryArguments{}
 	for i, t := range stmtMeta.inTypes {
-		plen, err := buf.getInt32()
+		plen, err := buf.getUint32()
 		if err != nil {
 			return err
 		}
 		k := fmt.Sprint(i + 1)
-		if plen == -1 {
+		if int32(plen) == -1 {
 			// The argument is a NULL value.
 			qargs[k] = parser.DNull
 			continue
@@ -558,7 +558,7 @@ func (c *v3Conn) handleBind(buf *readBuffer) error {
 		qargs[k] = d
 	}
 
-	numColumns := int16(len(stmt.Columns))
+	numColumns := uint16(len(stmt.Columns))
 	columnFormatCodes := make([]formatCode, numColumns)
 
 	// From the docs on number of result-column format codes to bind:
@@ -568,7 +568,7 @@ func (c *v3Conn) handleBind(buf *readBuffer) error {
 	// (if any); or it can equal the actual number of result columns of the
 	// query.
 	// http://www.postgresql.org/docs/current/static/protocol-message-formats.html
-	numColumnFormatCodes, err := buf.getInt16()
+	numColumnFormatCodes, err := buf.getUint16()
 	if err != nil {
 		return err
 	}
@@ -576,7 +576,7 @@ func (c *v3Conn) handleBind(buf *readBuffer) error {
 	case 0:
 	case 1:
 		// Read one code and apply it to every column.
-		c, err := buf.getInt16()
+		c, err := buf.getUint16()
 		if err != nil {
 			return err
 		}
@@ -587,7 +587,7 @@ func (c *v3Conn) handleBind(buf *readBuffer) error {
 	case numColumns:
 		// Read one format code for each column and apply it to that column.
 		for i := range columnFormatCodes {
-			c, err := buf.getInt16()
+			c, err := buf.getUint16()
 			if err != nil {
 				return err
 			}
@@ -613,7 +613,7 @@ func (c *v3Conn) handleExecute(ctx context.Context, buf *readBuffer) error {
 	if !ok {
 		return c.sendInternalError(fmt.Sprintf("unknown portal %q", portalName))
 	}
-	limit, err := buf.getInt32()
+	limit, err := buf.getUint32()
 	if err != nil {
 		return err
 	}
@@ -625,7 +625,7 @@ func (c *v3Conn) handleExecute(ctx context.Context, buf *readBuffer) error {
 		Values: portal.Qargs,
 	}
 
-	return c.executeStatements(ctx, stmt.Query, &pinfo, portalMeta.outFormats, false, limit)
+	return c.executeStatements(ctx, stmt.Query, &pinfo, portalMeta.outFormats, false, int(limit))
 }
 
 func (c *v3Conn) executeStatements(
@@ -634,7 +634,7 @@ func (c *v3Conn) executeStatements(
 	pinfo *parser.PlaceholderInfo,
 	formatCodes []formatCode,
 	sendDescription bool,
-	limit int32,
+	limit int,
 ) error {
 	tracing.AnnotateTrace()
 	results := c.executor.ExecuteStatements(ctx, c.session, stmts, pinfo)
@@ -708,7 +708,7 @@ func (c *v3Conn) sendErrorWithCode(errCode string, errCtx sqlbase.SrcCtx, errToS
 	return c.wr.Flush()
 }
 
-func (c *v3Conn) sendResponse(results sql.ResultList, formatCodes []formatCode, sendDescription bool, limit int32) error {
+func (c *v3Conn) sendResponse(results sql.ResultList, formatCodes []formatCode, sendDescription bool, limit int) error {
 	if len(results) == 0 {
 		return c.sendCommandComplete(nil)
 	}
@@ -719,7 +719,7 @@ func (c *v3Conn) sendResponse(results sql.ResultList, formatCodes []formatCode, 
 			}
 			break
 		}
-		if limit != 0 && len(result.Rows) > int(limit) {
+		if limit != 0 && len(result.Rows) > limit {
 			if err := c.sendInternalError(fmt.Sprintf("execute row count limits not supported: %d of %d", limit, len(result.Rows))); err != nil {
 				return err
 			}
